requirements.txt
flask==2.3.3
langchain==0.0.335
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.17.1
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3 